update to fit new caller
