+++ experimental/sdk/webpubsub-socketio-extension/src/common/azure-api/operation-spec.ts
