+++ experimental/sdk/webpubsub-socketio-extension/src/common/azure-api/parameters.ts
